+++ src/tquality_core/models/__init__.py
@@ -1,0 +1,8 @@
+from tquality_core.models.config import BaseConfig, WaiterConfig
+from tquality_core.models.jsonc_settings_source import JsoncConfigSettingsSource
+
+__all__ = [
+    "BaseConfig",
+    "JsoncConfigSettingsSource",
+    "WaiterConfig",
+]